Adds module structure
