+++ index.html
@@ -7,8 +7,7 @@
     <link rel="stylesheet" href="css/style.css">
     <link rel="shortcut icon" href="favicon.ico" type="image/x-icon">
     <title>Кекстаграм</title>
-    <script defer src="js/functions.js"></script>
-    <script defer src="js/main.js"></script>
+    <script src="js/main.js" type="module"></script>
 </head>
 
 <body>
